+++ CHANGELOG.md
@@ -832,7 +832,7 @@ popup notification opens up.
 - Fix bug where provider menu did not allow switching to custom network from a custom network.
 - Sending a transaction from within Badger no longer triggers a popup.
 - The ability to build without livereload features (such as for production) can be enabled with the gulp --disableLiveReload flag.
-- Fix Ethereum JSON RPC Filters bug.
+- Fix Bitcoin Cash JSON RPC Filters bug.
 
 ## 2.10.0 2016-08-29
 
@@ -868,7 +868,7 @@ popup notification opens up.
 - Add a form for Coinbase to specify amount to buy
 - Fix various typos.
 - Make dapp-metamask connection more reliable
-- Remove Ethereum Classic from provider menu.
+- Remove Bitcoin Cash Classic from provider menu.
 
 ## 2.7.3 2016-07-29
 
@@ -876,7 +876,7 @@ popup notification opens up.
 
 ## 2.7.2 2016-07-29
 
-- Add Ethereum Classic to provider menu
+- Add Bitcoin Cash Classic to provider menu
 - Fix bug where host store would fail to receive updates.
 
 ## 2.7.1 2016-07-27